Discremenate of different stages of the project
