+++ DataLink_Parser/backend/src/main/java/com/backend/DataParser.java
@@ -1,0 +1,5 @@
+package com.backend;
+
+public class DataParser {
+    
+}
