+++ DataLink_Parser/frontend/src/main/java/com/frontend/App.java
@@ -1,0 +1,13 @@
+package com.frontend;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
