+++ DataLink_Parser/simulator/src/main/java/com/simulator/Client.java
@@ -1,0 +1,34 @@
+package com.simulator;
+
+import java.io.*;
+import java.net.*;
+
+public class Client {
+    private Socket clientSocket;
+    private PrintWriter out;
+    //private BufferedReader in;
+
+    public void startConnection(String ip, int port){
+        try {
+        clientSocket = new Socket(ip, port);
+        out = new PrintWriter(clientSocket.getOutputStream(), true);
+        //in = new BufferedReader(new InputStreamReader(clientSocket.getInputStream()));
+        }
+        catch (ConnectException e){
+            System.out.println("\nERROR: The server has not been started yet!\n");
+        }
+        catch (IOException e){
+            System.out.println(e);
+        }
+    }
+
+    public void sendMessage(String msg) throws IOException{
+        out.println(msg);
+    }
+
+    public void stopConnection() throws IOException{
+        //in.close();
+        out.close();
+        clientSocket.close();
+    }
+}
