+++ DataLink_Parser/simulator/src/main/java/com/simulator/binary_data/data_1.txt
@@ -1,0 +1,4 @@
+1011010101000110000111101000101101100101101101100010100000101110
+1111101100111000101101111101010001000101110100000001011011111010
+0111000000101101101110011001001100001101100110001001100000111111
+1001101110111111111110000101111000001111000100110100100010010110
