+++ DataLink_Parser/simulator/src/main/java/com/simulator/binary_data/data_3.txt
