+++ DataLink_Parser/simulator/src/main/java/com/simulator/exceptions/EmptyFileException.java
@@ -1,0 +1,8 @@
+package com.simulator.exceptions;
+
+public class EmptyFileException extends Exception{
+    public EmptyFileException(String s)
+    {
+        super(s);
+    }
+}
