+++ DataLink_Parser/simulator/src/main/java/com/simulator/exceptions/NotSixtyFourBitDataException.java
@@ -1,0 +1,8 @@
+package com.simulator.exceptions;
+
+public class NotSixtyFourBitDataException extends Exception {
+    public NotSixtyFourBitDataException(String s)
+    {
+        super(s);
+    }
+}
