+++ DataLink_Parser/simulator/src/main/java/com/simulator/exceptions/NullLineException.java
@@ -1,0 +1,8 @@
+package com.simulator.exceptions;
+
+public class NullLineException extends Exception {
+    public NullLineException(String s)
+    {
+        super(s);
+    }
+}
