--- project/src/main/java/com/havelsan/App.java
@@ -1,13 +1,0 @@
-package com.havelsan;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
