--- project/src/main/java/com/havelsan/backend/DataParser.java
@@ -1,5 +1,0 @@
-package com.havelsan.backend;
-
-public class DataParser {
-    
-}
