--- project/src/main/java/com/havelsan/backend/StartServer.java
@@ -1,14 +1,0 @@
-package com.havelsan.backend;
-
-import java.io.*;
-
-public class StartServer {
-    public static void main(String[] args) throws IOException {
-
-        Server server = new Server();
-    
-        server.start(5000);
-        while(server.listen().equals("0000000000000000000000000000000000000000000000000000000000000000") == false);
-        server.stop();
-    }
-}
