--- project/src/main/java/com/havelsan/exceptions/EmptyFileException.java
@@ -1,8 +1,0 @@
-package com.havelsan.exceptions;
-
-public class EmptyFileException extends Exception{
-    public EmptyFileException(String s)
-    {
-        super(s);
-    }
-}
