--- project/src/main/java/com/havelsan/exceptions/NotSixtyFourBitDataException.java
@@ -1,8 +1,0 @@
-package com.havelsan.exceptions;
-
-public class NotSixtyFourBitDataException extends Exception {
-    public NotSixtyFourBitDataException(String s)
-    {
-        super(s);
-    }
-}
