--- project/src/main/java/com/havelsan/exceptions/NullLineException.java
@@ -1,8 +1,0 @@
-package com.havelsan.exceptions;
-
-public class NullLineException extends Exception {
-    public NullLineException(String s)
-    {
-        super(s);
-    }
-}
